Implemented component that deactivates objects when they're outside loaded chunks

+++ include/interaction.h
@@ -11,8 +11,8 @@ struct HealthState
 };
 
 // The width of a health bar is the entity's max health divided by this number, times four (since fillrects need to be a multiple of 4)
-constexpr size_t health_per_pixel = 4;
-constexpr size_t health_bar_height = 4;
+constexpr int health_per_pixel = 4;
+constexpr int health_bar_height = 4;
 
 Entity *findClosestEntity(Vec3 pos, archetype_t archetype, float maxDist, float *foundDist, Vec3 foundPos);
 int take_damage(Entity* entity, HealthState& health_state, int damage);
